[FEATURE] Implement OD store/restore support in API layer
On access to objects 0x1010 or 0x1011 the OD is stored to non-volatile
memory or factory defaults are restored respectively. If valid OD archive
for communication profile is present, storable configuration parameters
from tEplApiInitParam structure (like cycle length) are not used. Also the
CDC file is not loaded in that case.
The actual non-volatile memory access is implemented in the target-specific
module EplTgtObdArc.

Warnings:
1. CDC file loading should be split in separate object index ranges:
   communication profile 0x1000-0x1FFF and the rest. The latter one must
   not be loaded in case of NMT Reset Communication.
2. CFM support: Storage of variable size object entries like CN CDCs in
   object index 0x1F22 is currently not supported by this implementation.
   So leave EPL_OBD_USE_STORE_RESTORE disabled if CFM is enabled.

Additionally, this commit fixes a critical compiler warning in
EplApiUpdateObd() on 64 bit Linux machines regarding comparison against
~0UL. It needs a typecast to DWORD.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>